dawn::wire::server: Add Known<T> for ID + ObjectData<T>
Then when an ObjectId has a known type it points to, pass a Known<T>
instead of the raw id to the doers. This saves some manual error
handling.

Use Known<T>::AsHandle() when possible to get an ObjectHandle to reduce
the verbosity of doers.

Removes the redundant setting of obj->generation after a call to
KnownObjects::Allocate as that call already sets the generation.

Type a couple uint32_t with ObjectId and ObjectGeneration when
applicable.

Also rename `resultData` to more explicit names in doers.

Bug: dawn:1821
Change-Id: I7b1137b1a16e3319ae54062038c33a279ff16b2b
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/134640
Kokoro: Kokoro <noreply+kokoro@google.com>
Reviewed-by: Austin Eng <enga@chromium.org>
Commit-Queue: Corentin Wallez <cwallez@chromium.org>