[ir] Convert block instructions to a list.
This Cl moves the block instructions from a vector to a list. There will
be a lot of inserting and shuffling of these instructions when doing
transforms so this makes updating and removing simpler.

The block has several instructions added for working with the
instruction list, so, in effect, the block is at a high level an
instruction list container. There is a `begin` and `end` iterator on
block which iterates over the instructions.

Bug: tint:1718
Change-Id: I068801b390ca1e99baf22afb9712c12990c7f36d
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/135181
Reviewed-by: Ben Clayton <bclayton@google.com>
Kokoro: Kokoro <noreply+kokoro@google.com>
Commit-Queue: Dan Sinclair <dsinclair@chromium.org>
Reviewed-by: James Price <jrprice@google.com>