[ir] Add accessors
This CL adds member and index accessors to the IR. This is done with a
new `access` instruction which takes a pointer and set of indices and
returns a pointer to the last indexed item.

Swizzles are handled separately. A single item swizzle works as an `access` to the given value. Multi element swizzles produce a `swizzle` element which takes a value and mask and produces a new value.

This CL adds supports for loading index and member accessors from `var`s along with single element swizzles. `let` support and multi element swizzles will be done in a followup.

Bug: tint:1913
Change-Id: I3822a1ba94edb489dde5bf61a0c6ed0156cec726
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/135040
Commit-Queue: James Price <jrprice@google.com>
Reviewed-by: James Price <jrprice@google.com>
Reviewed-by: Ben Clayton <bclayton@google.com>
Kokoro: Kokoro <noreply+kokoro@google.com>