[tint][ir] Add MultiInBlock that derives from Block
IncomingBranches and BlockParams are only used to track branches from
two different blocks. Create a new MultiInBlock class that derives
from Block, and moves the IncomingBranches and BlockParams to this
new derived class. Rename IncomingBranches to InboundSiblingBranches
to signify that the ControlInstruction is not included in this list.

The only places that MultiInBlock are used are on the merge target of
If, Loop and Switch, and the Body and Continuing blocks of a Loop.
The rest of the blocks have a single incoming branch, which is
implicitly known by the flow-control instruction.

Bug: tint:1718
Change-Id: I49b54950076412815457b1cc2566368e55b19bd2
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/136280
Commit-Queue: Ben Clayton <bclayton@google.com>
Reviewed-by: Dan Sinclair <dsinclair@chromium.org>
Kokoro: Kokoro <noreply+kokoro@google.com>