[ir] Unconst most of the IR.
The IR is designed to be mutated. There are several `const` methods in
both the IR itself and the disassembler and validator. These methods are
unweildy in the face fo the IR being mutable.

This CL removes the `const` from most of the IR methods. The
`type::Type`, `constant::Value` and `Slice<T* const>` remain as const
due to ownership constraints.

Bug: tint:1959
Change-Id: Idbe41bc6bcc06621d9a6bea60cb8370c121a2f33
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/136621
Commit-Queue: Dan Sinclair <dsinclair@chromium.org>
Kokoro: Kokoro <noreply+kokoro@google.com>
Reviewed-by: Ben Clayton <bclayton@google.com>