[tint][transform][hlsl] Fix bad materialization
As reported in tint:1963, the PromoteSideEffectsToDecl transform could
hoist a constant expression to an implicitly typed 'let' declaration,
resulting in the expression type materializing early and resolve to a
different type.

Constant expressions don't need to be hoisted, but it seems generally
safer to continue emitting these, but with an explicit type on the
'let' declaration. This is what this CL does.

In doing this, I uncovered a second bug where the DecomposeMemoryAccess
transform would create a new a struct for the result of the
atomicCompareExchangeWeak() intrinsic, which would fail to assign to
the new explicitly typed lets. The fix for this is simple - don't
generate a new struct, but instead use the (fairly new) internal struct
declarations for these. This keeps the resolver happy, and actually
reduces a bunch of spew from the HLSL output.

Fixed: tint:1963
Change-Id: I271a2130dc303da3b9f11876ec5c6e8b98a16ee7
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/137100
Kokoro: Kokoro <noreply+kokoro@google.com>
Reviewed-by: James Price <jrprice@google.com>
Commit-Queue: Ben Clayton <bclayton@google.com>