[ir] Update `Value` parameters.
This CL updates the various instructions to allow the provided `Value`s
to be `nullptr`. These will be checked by the validator to ensure they
do not move through to the backends.

The `AddOperand` and `AddOperands` methods are updated to take the
operand position. The provided position is then validated against
the operand list to validate that it is the next appended item. This
verifies that the operands are added at the expected indices so any call
to `SetOperand` will work as expected.

Bug: tint:1718
Change-Id: I2d20f1d8a18da80a70c45e49ae8598fe477eb826
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/137360
Reviewed-by: James Price <jrprice@google.com>
Commit-Queue: Ben Clayton <bclayton@google.com>
Kokoro: Kokoro <noreply+kokoro@google.com>
Reviewed-by: Ben Clayton <bclayton@google.com>