Disable "use_dxc" flag when using D3D11 device
When using the D3D11 device, forcibly unset the use_dxc flag to avoid
issues with checking if the flag is set later, such as when dumping
shaders. Note that this mirrors how the D3D12 device will forcibly set
this flag to false if use_dxc is set, but the DXC dlls are not found.

This fixes an assert when use_dxc and dump_shaders are enabled on D3D11.

Also, opportunistically make dump_shader for DXC also dump the
compilation flags, as we do for FXC.

Change-Id: I74cab9c1e74aa177637dd51267b40cce38b80d0c
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/137740
Commit-Queue: Antonio Maiorano <amaiorano@google.com>
Kokoro: Kokoro <noreply+kokoro@google.com>
Reviewed-by: Austin Eng <enga@chromium.org>