[ir][spirv-writer] Add MergeReturn transform
Each return instruction is replaced with instructions that set a flag
to indicate the function is returning, capture the return value, and
exit from the current block to the nearest merge block. The code in
the merge block is then conditionalized based on that flag. A single
return instructions is then added to the final merge block of the
function.

This is needed to ensure that we adhere to SPIR-V's uniformity
requirements.

Use it in the SPIR-V writer.

Bug: tint:1906
Change-Id: I80edef751acc8e807a78e55549244a61726b462e
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/137340
Reviewed-by: Dan Sinclair <dsinclair@chromium.org>
Reviewed-by: Ben Clayton <bclayton@google.com>
Kokoro: Kokoro <noreply+kokoro@google.com>
Commit-Queue: James Price <jrprice@google.com>