[tint][ir] Refactor ToProgram
Rework the implementation so that all instructions are visited once.
Those that produce expression values are stored as either an AST
expression node (if inlined in the single place of usage), or as the
name of a let.

This approach has the benefit that we do not use recursion for
expression evaluation, substantially reducing the risk of stack
overflow.

Change-Id: I1fc0787a2e69f2a999227fe9198d5bfc442aff14
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/137165
Reviewed-by: Dan Sinclair <dsinclair@chromium.org>
Commit-Queue: Ben Clayton <bclayton@google.com>
Kokoro: Kokoro <noreply+kokoro@google.com>