[tint][ir] Reformat Block & ControlInstruction disassembly
Wrap the content of ControlInstruction in braces.
Put the block comment on the block line itself.

This is done to help visualize the content of the ControlInstruction from the instructions that are below the instruction.
By placing the block comment at the end of the block declaration, we reduce the amount of vertical whitespace, which can make indentation tracking more difficult.

Change-Id: Idba4f3c1c93d2d5e47a6395c9baafc06bcd304e9
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/138060
Reviewed-by: Dan Sinclair <dsinclair@chromium.org>
Kokoro: Kokoro <noreply+kokoro@google.com>