[msl] Simplify `EmitType` in MSL Generator.
Currently the `EmitType` method takes a `name` and a `name_printed`
flag. In the case of a pointer the name is emitted. It is always emitted
at the end of the pointer and most places which use the name have logic
to then emit it if not a pointer.

This CL removes the `name` and `name_printed` flags and just prints what
is needed after the `EmitType` call.

Change-Id: Idf44e314499fa5c61d6344688cb995afa1042817
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/138885
Reviewed-by: Ben Clayton <bclayton@google.com>
Commit-Queue: Dan Sinclair <dsinclair@chromium.org>
Kokoro: Kokoro <noreply+kokoro@google.com>