Merges AttachmentState and AttachmentStateBlueprint.
Maintaining the blueprint type makes the cache significantly more
complex. This merge tries to get the best of both worlds, allowing the
AttachmentState to be created on the stack when trying to check for
cache hits, but also allowing a copy of it to a heap allocated version
when needed. This change also addresses an existing unintended behavior
where rehashing of the AttachmentState cache would result in calls to
AttachmentStateBlueprint::HashFunc which recomputes the hash instead of
using a cached version like all other CachedObjects.

Downside is that now when we create the stack AttachmentState, we
additionally need to allocate the uint64_t for the refcount, and a
pointer for the Device.

Bug: dawn:1769
Change-Id: I0f0688c54b45f769791dfb228626a1d286247e02
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/139504
Commit-Queue: Loko Kung <lokokung@google.com>
Reviewed-by: Austin Eng <enga@chromium.org>
Kokoro: Kokoro <noreply+kokoro@google.com>
Reviewed-by: Corentin Wallez <cwallez@chromium.org>