[ir][validation] Walk through if/switch/loop in the validator
Add support for walking through the blocks of if/switch and loop now
that the structure has settled. Fixup a few tests along the way that are
broken due to failed validation.

This updates the disassembler to assume that an if always has a true
block and a loop always has a body. This is then validated by the generic block code that a terminator is required in a block.

Bug: tint:1952
Change-Id: I8e18e8b314b1b113602fa279143b96d870706690
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/139505
Reviewed-by: Ben Clayton <bclayton@google.com>
Kokoro: Kokoro <noreply+kokoro@google.com>
Commit-Queue: Dan Sinclair <dsinclair@chromium.org>