[ir] Hookup IR to test runner
This CL adds a `--use-ir` flag to the test runner which will set the
expectations to `expected.ir.<type>` and pass the `--use-ir` to the tint
program.

The tint program is updated to pass the use_ir flag through to the MSL
backend when enabled.

The SPIR-V generator is updated to set a generator version number of `1`
into the new IR generator to make it possible to know which version of
the generator created the SPIR-V.

Bug: tint:1718
Change-Id: If7584a919a916bf6fe6fdf1158bf635e15150e97
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/139880
Reviewed-by: Ben Clayton <bclayton@google.com>
Reviewed-by: James Price <jrprice@google.com>
Commit-Queue: Dan Sinclair <dsinclair@chromium.org>
Kokoro: Kokoro <noreply+kokoro@google.com>