d3d11: Fix firstIndexOffsetShaderRegister conflict
Currently its value is assigned to kReservedConstantBufferSlot, which
is in fact 13. However, this is an available binding number in WGSL to
be used by clients. So this can lead to conflicts potentially. To fix
it, this CL first puts it into an unused group and then remaps it to
space 0, and slot kReservedConstantBufferSlot.

Bug: dawn:1905
Bug: dawn:1705

Change-Id: I0ca837d3ccab06a753af9203b489087e80c46384
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/140142
Reviewed-by: Austin Eng <enga@chromium.org>
Kokoro: Kokoro <noreply+kokoro@google.com>
Commit-Queue: Jie A Chen <jie.a.chen@intel.com>