[tint][ir] Add LoadVectorElement / StoreVectorElement
These new instructions must now be used to load and store a vector
element. It is now a validation error to obtain a pointer to a
vector member.

This is done to match the limitations of MSL and WGSL.

Change-Id: I7a0fa7287ecfc7cd441f9cda432df1134c4bc45a
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/139924
Commit-Queue: Ben Clayton <bclayton@google.com>
Kokoro: Ben Clayton <bclayton@google.com>
Reviewed-by: Dan Sinclair <dsinclair@chromium.org>
Reviewed-by: James Price <jrprice@google.com>