[ir] Implement demote-to-helper transform
Find fragment shaders that discard, replace their discard instructions
with a global flag, and conditionalize stores to host-visible address
spaces.

This requires two passes through the module, the first to find the
entry points that (transitively) contain a discard, and the second to
process functions that belong to a shader that may discard.

Introduces a TerminateInvocation instruction which is a block
terminator. A conditional TerminateInvocation instruction is inserted
before every return instruction in the entry point.

Bug: tint:1718

Change-Id: I2ade9c014190b1270a83a37877af310cc9a6db24
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/140886
Commit-Queue: James Price <jrprice@google.com>
Reviewed-by: Dan Sinclair <dsinclair@chromium.org>
Reviewed-by: Ben Clayton <bclayton@google.com>
Kokoro: Kokoro <noreply+kokoro@google.com>