[ir][spirv-writer] Polyfill the `select()` builtin
The argument order of OpSelect is different, and the overload that
produces a vector with a scalar condition is not supported until
SPIR-V 1.4.

Adds a BuiltinPolyfillSpirv transform for these polyfills. Use that
transform in the backend, and map the `spirv.select` intrinsic to the
`OpSelect` instruction.

Bug: tint:1906
Change-Id: I1d42d41ccca4a5e7af4140514afa02a12a72bf28
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/140889
Auto-Submit: James Price <jrprice@google.com>
Commit-Queue: James Price <jrprice@google.com>
Reviewed-by: Dan Sinclair <dsinclair@chromium.org>
Kokoro: Kokoro <noreply+kokoro@google.com>