[ir][validation] Move Validator to header file
This Cl moves the Validator implementation out of an anonymous namespace
in the cc file to the header file. This will allow backends to subclass
the validator and create custom validation.

Currently none of the methods are virtual, they will move over as we
find a need to override them in a backend.

Bug: tint:1952
Change-Id: If43baf943742457db597f9add3ca6b4b80b42843
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/141241
Reviewed-by: Ben Clayton <bclayton@google.com>
Kokoro: Kokoro <noreply+kokoro@google.com>
Commit-Queue: Dan Sinclair <dsinclair@chromium.org>