[shuffle] Move msl writers to new structure.
This CL moves the `msl/ast_writer` to `msl/writer/ast_printer/` and
renames the `GeneratorImpl` to `ASTPrinter`. The `msl/writer` is moved
to `msl/writer/printer` and the `GeneratorImplIr` is renamed to
`Printer`. The `Generator` is moved to `msl/writer` and renamed to
`Writer`. The namespace is updatd to `tint::msl::writer`.

Bug: tint:1988
Change-Id: Ic09b1a7b536497b4056b4d8fd2933a08ccf0f1be
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/143080
Reviewed-by: Ben Clayton <bclayton@google.com>
Kokoro: Kokoro <noreply+kokoro@google.com>
Commit-Queue: Dan Sinclair <dsinclair@chromium.org>