Split Vulkan command buffers on compute pass
This is a further generalization of the workaround added for
dawn:1564. It splits the command buffer on Qualcomm GPUs any time
a compute pass follows a render pass in a single command buffer.
This feels like a farily heavy-handed workaround, but where
previously it seemed like the crash only happened if the compute
pass used attachements from the render pass, I'm now seeing some
scenarios where the compute pass shares no resources with the
render pass and still fails unless the command buffer is split.

Fortunately the broadening of scope allows the code to be simpler
and removes the need to track and check render pass attachements.

Bug: dawn:1897
Change-Id: I2c511a22e745ee5804d46337cb52483c009a200e
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/144780
Reviewed-by: Kai Ninomiya <kainino@chromium.org>
Commit-Queue: Brandon Jones <bajones@chromium.org>
Kokoro: Kokoro <noreply+kokoro@google.com>