Only unmap staging buffer before CopyFromStagingToBuffer() on D3D11
This patch moves the unmap() call on staging buffer to D3D11 backend
because only D3D11 backend requires to do so.

Note that if we keep unmap() before CopyFromStagingToBuffer() on all
backends, it will be difficult to share the staging buffer for
mappedAtCreation among different buffers because it is possible that
when we call unmap() on one buffer, another buffer using same staging
buffer internally is still mapped and waiting for data to be written
in. Because of this reason it seems we cannot easily reuse the
staging buffer for mappedAtCreation on D3D11 backend.

Bug: dawn:828, dawn:1772
Change-Id: Ibe92df69ee459669630af4d5898fd4d899d0f63a
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/145180
Kokoro: Kokoro <noreply+kokoro@google.com>
Reviewed-by: Austin Eng <enga@chromium.org>
Reviewed-by: Corentin Wallez <cwallez@chromium.org>
Commit-Queue: Jiawei Shao <jiawei.shao@intel.com>