[tint] Rework the intrinsic table to use indices
These are smaller than 64-bit pointers so use less size in the binary,
and will be friendlier on the cache.

Fix the intrinsic table LUT compression. We were storing pointers in the
keys, which prevented a substantial amount of compression.

Migrate 'lut' to using Go generics.
Simplfies the code, and removes the need for the 'list' package.

Change-Id: I7c12bc094c4188cbf4946bb32fe40522f5cdb6d0
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/145521
Reviewed-by: James Price <jrprice@google.com>
Kokoro: Kokoro <noreply+kokoro@google.com>
Reviewed-by: Dan Sinclair <dsinclair@chromium.org>