Fix GLTextureWrappingTests: create a second GL device.
It turns out that the existing tests still pass even if the relevant ANGLE extension is not requested in ContextEGL. The reason is that the tests were using the test's native::opengl::Device to create a GL texture, wrap it and send it back to the same device (and same EGL context). This works even without the extension, since the GL texture IDs match.

A better idea (stolen from VulkanImageWrappingTests) is to create a second native::opengl::Device and create textures in its context. That way, the texture IDs will be unknown to the main Device unless the extension is requested.

Bug: chromium:1414566
Change-Id: I134a5a8cb3724c8a6ccb65a26e328ea75b763c35
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/145560
Reviewed-by: Austin Eng <enga@chromium.org>
Kokoro: Kokoro <noreply+kokoro@google.com>
Commit-Queue: Stephen White <senorblanco@chromium.org>