[msl] Add support for chromium_experimental_subgroups
Use a transform to replace calls to `subgroupBallot()` with a helper
function that masks out the unused bits, as well as converting the
64-bit output from MSL's `simd_active_threads_mask()` function to a
vec4u.

Adds the ability to change the target MSL version used for validation,
which is automatically bumped to MSL 2.1 if the subgroups extension is
used.

Bug: tint:2000
Change-Id: I3161b249acd90378be52b18f6325c82fd9dc80b2
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/146460
Commit-Queue: James Price <jrprice@google.com>
Reviewed-by: Ben Clayton <bclayton@google.com>
Kokoro: Kokoro <noreply+kokoro@google.com>