Remove entries from `build_overrides/tint.gni`
The build_override file was setting the overrides to the same value as
the defaults. The existence of the entries in build_overrides was also
causing them to not be declared as arguments in the
tint_overrides_with_defaults.gni file. This then means that anything
that was in build_overrides/tint.gni could not be overridden with gn
args settings.

Change-Id: I0552958acec57aaad7cdcea7173321b60685613d
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/147580
Kokoro: Kokoro <noreply+kokoro@google.com>
Reviewed-by: Ben Clayton <bclayton@google.com>
Reviewed-by: James Price <jrprice@google.com>
Commit-Queue: Dan Sinclair <dsinclair@chromium.org>