[eval] Remove template param from TransformBinaryDifferingArityElements
All of the callers use the same signature for the lambda function. This
CL defines an `BinaryTransfrom` std::function which is used instead of
the template parameter.

There are 65 separate copies of this function in the current chrome canary. One for each template usage. Each of those copies takes .25kb of space in the binary for a total of ~16k.

Change-Id: Ifeebe3710f119ab395290fed7906a520c87815f1
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/148280
Reviewed-by: Ben Clayton <bclayton@google.com>
Reviewed-by: Antonio Maiorano <amaiorano@google.com>
Kokoro: Kokoro <noreply+kokoro@google.com>
Commit-Queue: Dan Sinclair <dsinclair@chromium.org>