Add the validation for PixelLocalStorage.
The AttachmentState is modified to contain the PLS state as it is part
of the compatibility between a render pass and a render pipeline.

Minimal fixes to dawn.json for GPU sizes.

Adds tests for the added validation.

Bugs found with tests:
 - The AttachmentState should different between no PLS and empty PLS.
 - Missing usage tracking.
 - Attachment state blueprint init didn't copy mHasPLS
 - Pipeline layout hash/equality wasn't updated for PLS.
 - webgpu_absl_format didn't handle no PLS vs. empty PLS.

Bug: dawn:1704
Change-Id: If56bf8748c29d13f8d2ee1290b79e039fffc790d
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/147500
Reviewed-by: Austin Eng <enga@chromium.org>
Reviewed-by: Quyen Le <lehoangquyen@chromium.org>
Kokoro: Kokoro <noreply+kokoro@google.com>
Commit-Queue: Corentin Wallez <cwallez@chromium.org>