dawn: add dawn::Platform hook for querying feature flags
And use it to query "WebGPUUseDXC". This is important because the moment
the feature flag is queried in Chrome is when the user is made part of
the study. Since we normally want only users using WebGPU to be in the
study, we need this query to happen from within Dawn.

Bug: chromium:1474897
Change-Id: I2adad2b3a256fad5c51080b80b348c0ffd5a914d
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/153181
Kokoro: Kokoro <noreply+kokoro@google.com>
Commit-Queue: Antonio Maiorano <amaiorano@google.com>
Reviewed-by: Corentin Wallez <cwallez@chromium.org>