OpenGL: fix multiplanar external textures.
There are two problems with multiplanar textures in the OpenGL backend.

1) When the the multiplanar external texture transform runs, it splits
a texture_external variable into two texture_2d variables. The
Resolver then creates new texture/sampler pairs for the samplers with
which the plane1 texture is used, and new binding points for those
that bubble up to the global level. Since Dawn collects
sampler/texture pairs from the Inspector before the transforms run, it
doesn't know about these extra pairs, doesn't name them and doesn't
bind anything to them.

2) The multiplanar transform also inserts calls to
textureDimensions(), a texture-only builtin that doesn't require a
sampler. In the CombineSamplers transform, since OpenGLES doesn't have
the concept of a texture-only sampler2d, this necessitates the
creation of new binding points for the texture with a "placeholder"
sampler, which Dawn also doesn't know about, so it doesn't name them
or bind anything to them.

Fix #1 is on the Dawn side: when iterating over external texture
bindings, create a map from plane0 texture -> plane1 texture. When
iterating over the sampler/texture uses that the Tint inspector has
discovered, if the texture is an external texture, add a new pair with
the same sampler and the plane1 texture.

Fix for #2 is on the Tint side: if a function contains both a
texture-only reference and a texture/sampler reference to the same
texture, use the latter and remove the former. Texture-only builtins
won't use the sampler anyway. This is essentially an optimization, but
it's one we know will run on the functions generated by the
multiplanar transform, since it always calls textureDimensions() in a
function also containing the textureSample() call on the same texture.

Change-Id: I531ab09acdc47b435d72033aabd9eefa57f8f6e3
Bug: tint:1774
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/152660
Kokoro: Kokoro <noreply+kokoro@google.com>
Reviewed-by: Ben Clayton <bclayton@google.com>
Commit-Queue: Stephen White <senorblanco@chromium.org>