[ir] Fix let instructions in BlockDecoratedStructs
When the struct is not wrapped there are type mismatches between the
old struct and the new struct. Instead of generating an unnecessary
wrapper in these cases, just mark the original struct as a block. This
simplifies the transform logic at the cost of a `const_cast`, which
will be removed when we eventually un-const types.

Change-Id: Id60eb6e45768acc15f1f6dd76d761a3d657d37d2
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/153861
Reviewed-by: Ben Clayton <bclayton@google.com>
Commit-Queue: James Price <jrprice@google.com>
Kokoro: Kokoro <noreply+kokoro@google.com>