[tint][cmake] Rework targets for fuzzers
Instead of trying to juggle a plain target and a fuzzer-variant target
in the functions, import the generated targets twice - once with fuzzers
disabled and again with fuzzers enabled.
The second pass has a suffix on the tint targets, so we get our forked
fuzzer-enabled targets.

This fixes an issue I spotted where a number of fuzzer dependencies were
not getting compiled with `-fsanitize=fuzzer`, and so we were missing
coverage information for these.

Change-Id: I64ca703ab2ccbe3bee3c03e5298946843b2c4950
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/153781
Kokoro: Kokoro <noreply+kokoro@google.com>
Commit-Queue: Ben Clayton <bclayton@google.com>
Reviewed-by: Antonio Maiorano <amaiorano@google.com>