Lift buffer host mapped restrictions; test storage usage
This allows host mapped buffers to be used directly on the GPU.
It works for both UMA and non-UMA devices, though direct usage on
the GPU in the general case will not be efficient on non-UMA devices.
Profiling and choosing whether to make an intermediate copy is left
to the application to optimize.

Bug: dawn:2018
Change-Id: Ib153d4559ef3ed25cb9a9cdf60b181f69c2a879c
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/155461
Reviewed-by: Corentin Wallez <cwallez@chromium.org>
Kokoro: Kokoro <noreply+kokoro@google.com>
Commit-Queue: Austin Eng <enga@chromium.org>
Reviewed-by: Loko Kung <lokokung@google.com>