Move tint writer outputs into the SPIRV request.
This CL moves the options which are passed to the tint writer for the
SPIR-V backend out of the compilation callback and passes them directly
through the `SpirvCompilationRequest`. This should incur less copying of
the data in the request.

Change-Id: I1a91ec348001fc8d463fc424d51db60d4723d939
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/154100
Commit-Queue: dan sinclair <dsinclair@chromium.org>
Kokoro: Kokoro <noreply+kokoro@google.com>
Reviewed-by: Corentin Wallez <cwallez@chromium.org>
Reviewed-by: Ben Clayton <bclayton@google.com>