Move tint writer options into MSL request.
This CL moves the options which are passed to the tint writer for the
MSL backend out of the compilation callback and passes them directly
through the `MslCompilationRequest`. This should incur less copying of
the data in the request.

Change-Id: I36d4c3803cf88012a119df5e128e6e7072b026e6
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/154362
Reviewed-by: Ben Clayton <bclayton@google.com>
Reviewed-by: Corentin Wallez <cwallez@chromium.org>
Kokoro: Kokoro <noreply+kokoro@google.com>
Commit-Queue: dan sinclair <dsinclair@chromium.org>