Move tint writer options into GLSL request.
This CL moves the options which are passed to the tint writer for the
GLSL backend out of the compilation callback and passes them directly
through the `GlslCompilationRequest`. This should incur less copying of
the data in the request.

Change-Id: I070fb7a71bf6e27fbc208533df76acc608b47351
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/156363
Commit-Queue: dan sinclair <dsinclair@chromium.org>
Kokoro: Kokoro <noreply+kokoro@google.com>
Reviewed-by: Corentin Wallez <cwallez@chromium.org>
Reviewed-by: Ben Clayton <bclayton@google.com>