Explicitly handle NULL Vulkan debug message IDs
The Vulkan spec states that the pMessageIdName for a debug message may
be NULL. We were previously passing it into a LogMessage (which wraps an
std::ostream) unchecked, but passing a NULL into an ostream is undefined
behavior. This was likely leading to crashes on some devices.

This patch updates the logging behavior to check for NULL and take a
different path if encountered to avoid triggering undefined behavior.

Bug: chromium:1497136
Change-Id: I431539dc21f51fa48bb754f5bc9c85acf2ce8f04
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/159040
Reviewed-by: Austin Eng <enga@chromium.org>
Auto-Submit: Brandon Jones <bajones@chromium.org>
Commit-Queue: Austin Eng <enga@chromium.org>
Kokoro: Kokoro <noreply+kokoro@google.com>