dawn.node: Fix rejecting promises in async pipeline compilation.
Previously the TypeError would be thrown immediately on
createPipelinAsync when it should reject the exception. Implement this
by allowing the converter to retain the exception instead of throwing
it, so that it can be acquired to reject the promise.

Also adds a bunch of [[nodiscard]].

Bug: None
Change-Id: Ib77480345e7862017fa03f6fab13b904a3c2e9a6
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/159362
Kokoro: Kokoro <noreply+kokoro@google.com>
Commit-Queue: Corentin Wallez <cwallez@chromium.org>
Reviewed-by: Ben Clayton <bclayton@google.com>