dawn.node: Throw TypeError on non-finite floating points
This is to match the WebIDL rules that floating point values should be
finite, unless they are "unrestricted" (which is never the case in
WebGPU at the moment).

This fixes a bunch of overridable constant tests for NaNs and Inifinity.

Bug: None
Change-Id: I0482dc51fa607ee89394f7b9f10493168da368dc
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/159364
Reviewed-by: Ben Clayton <bclayton@google.com>
Reviewed-by: Kai Ninomiya <kainino@chromium.org>
Kokoro: Kokoro <noreply+kokoro@google.com>
Commit-Queue: Corentin Wallez <cwallez@chromium.org>