[tint][resolver] Standardise attribute diagnostics
Use '@blah', instead of 'blah attribute'.
This is less verbose, and is closer to the user code.

Also change '@X must only be used with @Y' to '@X can only be used with @Y' in attribute diagnostics.
This is to avoid interpretation that only '@X' and '@Y' can exist on an AST node.

Change-Id: Ied96f5344f5ab6efeee15fd52b63bbf3965c9f7d
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/159782
Commit-Queue: Ben Clayton <bclayton@google.com>
Kokoro: Kokoro <noreply+kokoro@google.com>
Reviewed-by: James Price <jrprice@google.com>