[tint][resolver] Fix @index validation, tweak diagnostics
Validator::EntryPoint() was returning early for IndexAttribute, which
was missing later validation. Replace the if-else chain with a Switch()
which makes early return impossible.

Also change the diagnostics for attributes that can only be applied to a
specific stage, so that the message states what stage it can be used in.
Knowing whether the usage is input or output is not that helpful.

Change-Id: I47352a006f45ad6421aa102697216206d0e2044d
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/159783
Kokoro: Kokoro <noreply+kokoro@google.com>
Reviewed-by: James Price <jrprice@google.com>
Commit-Queue: Ben Clayton <bclayton@google.com>