[tint][resolver] Fix dual source blending validation
The check for an entry point using a non-zero @location and non-zero
@index would only fire if the non-zero index came after the @location.

This now only validates on the entry point (as opposed to a standalone
struct), but given that this is non-spec'd this seems fine.

Rework the validation, and change the diagnostic message to be clearer
about what you did wrong.

Change-Id: I3e35b316814cb149c1f8f77f59e236463d87b2aa
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/160082
Commit-Queue: Ben Clayton <bclayton@google.com>
Kokoro: Kokoro <noreply+kokoro@google.com>
Reviewed-by: James Price <jrprice@google.com>