[tint][ir] Refactor Instruction result methods
Remove HasResults() and HasMultiResults(). These are virtual functions,
and are usually followed up with another virtual call to Result() or
Results(). This pattern is expensive and so should be discouraged.

Add an index to Operand() and Result() to keep these accessors
symmetrical.

Replace the Args() virtual methods with a virtual ArgsOperandOffset().
When we split methods into const and non-const flavors, this reduces the
number of duplicated methods as we can implement a const and non-const
Args() in the base class.

Change-Id: I7d0ecdb58746dcba90fc4631fead235474ccf1d3
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/161005
Reviewed-by: James Price <jrprice@google.com>
Reviewed-by: dan sinclair <dsinclair@chromium.org>
Kokoro: Kokoro <noreply+kokoro@google.com>