Mark SharedTextureMemory stable for DXGI and IOSurface
This feature is not supported over the wire, meaning that it is not
possible for Chromium renderer processes to use it. Hence it is not a
problem to mark it stable so that the Chromium GPU-side internal
implementation can assume that it is present. This will allow removing
usage of the WrapIOSurface() API from Chromium, which in turn will
allow cleanly optimizing the IOSurface implementation of
SharedTextureMemory.

Change-Id: I670491395e12fb6bc8b3afcc135eee3d8f2a44f6
Bug: chromium:1493854, dawn:2152
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/161640
Reviewed-by: Corentin Wallez <cwallez@chromium.org>
Commit-Queue: Corentin Wallez <cwallez@chromium.org>
Kokoro: Kokoro <noreply+kokoro@google.com>
Auto-Submit: Colin Blundell <blundell@chromium.org>