Metal: implement BufferMapExtendedUsages feature.
This feature allows MapRead or MapWrite to be combined with other
usages when creating buffer.

For example, a vertex buffer could be mapped with this feature enabled.
Unlike before when it could only be copied to from a mappable buffer.

Bug: dawn:2204
Change-Id: Ia4f177d27ea79d5fcb7f1670321a0e81a6de324e
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/161400
Commit-Queue: Quyen Le <lehoangquyen@chromium.org>
Reviewed-by: Corentin Wallez <cwallez@chromium.org>
Kokoro: Kokoro <noreply+kokoro@google.com>