[ir] Move use of IR decision higher in the stack.
This CL changes the tint writers to no longer except the `tint_use_ir`
flag. Instead, there is a version of `Generate` which accepts the IR and
one which accepts a Program. The caller is now responsible for passing
the correct params.

This decouples the WGSL reader from the IR writers which was required in
order to use the program to IR functionality.

Bug: tint:1718
Change-Id: If70a3f6c92a80adeee3094e485132ca33803817f
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/161742
Reviewed-by: Ben Clayton <bclayton@google.com>
Commit-Queue: dan sinclair <dsinclair@chromium.org>
Kokoro: Kokoro <noreply+kokoro@google.com>