Track OnSubmittedOnWorkDoneF futures using execution serials
This is needed on Vulkan because Vulkan cannot always make
system events from VkFences. Doing it on other backends
helps as well as an optimization. We don't need to create
a SystemEventPipe for *every* future, and instead create
them on demand when waiting is necessary.

Lazily creating system events then requires an extra level
of indirection to wrap them. `SystemEvent` does this
by storing an atomic boolean alongside an optional system
event sender/receiver pair.

Lastly, in order to avoid multiple vector allocations to convert
an array of TrackedFutureWaitInfo holding SystemEvent to arrays of
ready bools and SystemEventReceiver primitives which are passed to
the OS-specific waiting function, WaitAnySystemEvent is changed to
take a templated iterator so we can prepare the OS-specific handles
with fewer allocations.

This CL also removes the MustWaitUsingDevice concept from tracked
events since it will probably move to be a query on the device
instead of on each event. Mixed source waits are not supported yet,
and when they are, they will still want to reduce queue waits
to a single completion event. So, tracked events still need to
expose their queue/device instead of returning null when mixed
source waiting is enabled.

Fixed: dawn:2051
Fixed: dawn:2064
Bug: dawn:2150
Change-Id: I939d94c20b70c47013130d477029080c5240b804
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/156301
Reviewed-by: Loko Kung <lokokung@google.com>
Reviewed-by: Kai Ninomiya <kainino@chromium.org>
Kokoro: Kokoro <noreply+kokoro@google.com>
Commit-Queue: Austin Eng <enga@chromium.org>