[tint][wgsl][fuzz] Fix static init ordering bug
Because C++ makes no ordering guarantees, we were having some fuzzers
register themselves before the vector holding the registered fuzzers was
constructed. Upon construction, the vector would wipe away all the
registered fuzzers, and so we were only running those that were
registered after the vector was constructed.

Instead, use a static local to hold the vector. This is guaranteed to be
constructed the first time the function is called, avoiding the ordering
foot-gun.

Change-Id: I1f442e6a15c64f6125bd6866d8f8c00829f2055a
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/162303
Reviewed-by: dan sinclair <dsinclair@chromium.org>
Kokoro: Kokoro <noreply+kokoro@google.com>
Reviewed-by: James Price <jrprice@google.com>