Add tests of expected usages for SharedTextureMemory
This CL adds tests that SharedTextureMemory instances are created with
expected usages for single-planar and multiplanar textures. This will be
helpful for a followup where we extend the set of usages for multiplanar
textures based on the features available on the device.

Bug: 1493854
Change-Id: I88f5a5a998af48bbc223381d6d5438e681dcd616
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/161960
Commit-Queue: Colin Blundell <blundell@chromium.org>
Kokoro: Kokoro <noreply+kokoro@google.com>
Reviewed-by: Corentin Wallez <cwallez@chromium.org>