Mark SharedTextureMemory/SharedFence as supported on the wire
In Chromium we will need to use SharedTextureMemory/SharedFence on the
server-side for WebGPU devices. To do so we need to add the features on
such devices. However, doing so currently fails because the features
are not listed as wire-supported. This CL changes that to enable this
use case. Note that the features will still not actually be accessible
over the wire as they are tagged as "native" in dawn.json. Thus, this
CL will have the intended effect of allowing usage of these features
*only* from the server side in the context of client-requested Devices.

Change-Id: Ief7d1cc8b9d53843a2ccaaa9d1634c2d355c482d
Bug: chromium:1493854
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/162721
Reviewed-by: Corentin Wallez <cwallez@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Kokoro: Kokoro <noreply+kokoro@google.com>