[tint][cmake] Pass -Wno-gnu-zero-variadic-macro-arguments
This warning is raised with -Weverything when using a macro containing
`blah, ##__VA_ARGS__`.

All compilers we care about supports this already. It's also supported
as part of the standard with C++20.

Use this in the RUN_TRANSFORM() macro to remove a common argument.

Change-Id: If686bb53520de3b50649b9dcfda90659484dbf8d
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/163241
Reviewed-by: dan sinclair <dsinclair@chromium.org>
Commit-Queue: Ben Clayton <bclayton@google.com>
Kokoro: Kokoro <noreply+kokoro@google.com>