[tint][wgsl] Move LanguageFeature to a separate target
This target is special in that the GN build is entirely dependency-free.

This is to allow the target to be used outside of Dawn, and Chromium can
inspect the allowed features without having to create a Dawn device via
the wire.

Change-Id: I85ddfd76b897ae582fbffb734b4760872844190d
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/163320
Auto-Submit: Ben Clayton <bclayton@google.com>
Reviewed-by: Corentin Wallez <cwallez@chromium.org>
Kokoro: Kokoro <noreply+kokoro@google.com>
Commit-Queue: Ben Clayton <bclayton@google.com>