dawn::native: Do not record the backtrace on DAWN_TRY in Release
The use of this backtrace was mostly to help developers know the context
of WebGPU validation errors before we added DAWN_TRY_CONTEXT and
friends. Now it is mostly useful when developing and getting errors in
the backends, so disable the backtrace in release to save binary size.

Bug: None
Change-Id: I2dc667e25c68d92b8731680ba2210a8dae781903
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/165740
Kokoro: Kokoro <noreply+kokoro@google.com>
Commit-Queue: Corentin Wallez <cwallez@chromium.org>
Reviewed-by: Loko Kung <lokokung@google.com>
Reviewed-by: Austin Eng <enga@chromium.org>