spir-v ast writer: Fix order of dpdx and dpdy operands
We need to use a stable sort to preserve the order of the
Grad dpdx and dpdy operands.  Otherwise we're at the mercy
of the sort, since the two operands have the same mask.

Credit to amaiorano@ for finding the root cause and the supplying
the fix.

Change-Id: Ia352b1f2650980ac4e099fc223a0a84a93731679
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/166781
Commit-Queue: David Neto <dneto@google.com>
Kokoro: Kokoro <noreply+kokoro@google.com>
Auto-Submit: David Neto <dneto@google.com>
Reviewed-by: Antonio Maiorano <amaiorano@google.com>
Reviewed-by: James Price <jrprice@google.com>