Add test for external texture in different group from sampler
This CL adds a test for an external texture which exists in a different
bind group from the sampler that is used with the texture.

The test is currently disabled on Metal as it fails to build the binding
map correctly.

Bug: chromium:1515439
Change-Id: I6694974581863cab4ad594308ca1c2d5afd6409e
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/168200
Reviewed-by: Corentin Wallez <cwallez@chromium.org>
Kokoro: Kokoro <noreply+kokoro@google.com>
Commit-Queue: dan sinclair <dsinclair@chromium.org>