webgpu.h: Re-number enums, with defaulting: RenderPipelineDescriptor
Part 3 implements the following sub-structures of
RenderPipelineDescriptor. Defaulting is done in the constructor, which
(in most cases) is called by a bounce through the backend.

- PrimitiveState.topology: PrimitiveTopology = TriangleList
- PrimitiveState.frontFace: FrontFace = CCW
- PrimitiveState.cullMode: CullMode = None
- StencilFaceState.compare: CompareFunction = Always
- StencilFaceState.{fail,depthFail,pass}Op: StencilOperation = Keep
- BlendComponent.operation: BlendOperation = Add
- BlendComponent.srcFactor: BlendFactor = One
- BlendComponent.dstFactor: BlendFactor = Zero
- VertexBufferLayout.stepMode: VertexStepMode = Vertex

Spot tests are added for all of these. They won't necessarily catch if
the defaulting is _wrong_, but should crash if it is _missing_.

Bug: dawn:2224
Change-Id: I0e3029af0646ffcd58f8b8c99f7050206916ca36
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/166822
Kokoro: Kokoro <noreply+kokoro@google.com>
Reviewed-by: Loko Kung <lokokung@google.com>
Commit-Queue: Kai Ninomiya <kainino@chromium.org>