[tint][uniformity] Fix nested LValue expressions
When an LValue is a dereferenced pointer, we were cutting the analysis
short instead of descending into nested expressions that may contain
uniformity violations or affect the uniformity of the parent.

This CL adds an `is_dereferencing` parameter to
`ProcessLValueExpression()` to track whether we are currently
dereferencing, and moves the logic for partial pointers to the
handling of the root identifier expression. This unifies the logic in
the three different expressions that can cause derefences (unary *,
index accessors, and member accessors).

Bug: tint:2113, tint:2053
Fixed: tint:2139
Change-Id: Ia55b5345590c41ea1b162112ac2545c69a95ba75
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/169265
Commit-Queue: James Price <jrprice@google.com>
Reviewed-by: Antonio Maiorano <amaiorano@google.com>
Kokoro: Kokoro <noreply+kokoro@google.com>