Fix ASAN stack-use-after-scope in dawn_unittests
This unittest was using a stack variable to mock the mapped data.
The mapped data needs to live until the client's map write flushes
and completeds, because the map write mutates it.
This was a test bug, and not a problem in the implementation.

Fixed: chromium:1516605
Change-Id: I2c302ec7c78459084634d95f51ab997f1961a43f
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/169440
Kokoro: Kokoro <noreply+kokoro@google.com>
Reviewed-by: Corentin Wallez <cwallez@chromium.org>
Auto-Submit: Austin Eng <enga@chromium.org>
Commit-Queue: Corentin Wallez <cwallez@chromium.org>