[tint] Fix emission of module-scope struct initializers
Constant structure initializers need to be assigned directly to a
variable in HLSL. We were emitting these as `const` variables for both
module-scope and function-scope.

Module-scope `const` variables are not held in the program bitcode, but
are expected to be provided as a cbuffer binding, which dawn was not
providing.

Instead emit these always as global scope `static const` variables.

Add a dawn end2end test for `var<private>` variables initialized with a
`struct`

Bug: tint:2142
Change-Id: I03c7c94a1729dcfbc077bcef01495a9fbbff0290
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/169940
Commit-Queue: Ben Clayton <bclayton@google.com>
Kokoro: Kokoro <noreply+kokoro@google.com>
Reviewed-by: Antonio Maiorano <amaiorano@google.com>