Log error message if unable to open libEGL.
When attempting to load the libEGL library, we can retrieve the erorr
message for a failed load by providing a string. This CL updates the
call to provide the string and provides the context as to why the load
failed.

Change-Id: Iafa71d72960a415030750f83056354393592f4a1
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/170301
Commit-Queue: dan sinclair <dsinclair@chromium.org>
Reviewed-by: Gregg Tavares <gman@chromium.org>
Kokoro: Kokoro <noreply+kokoro@google.com>
Reviewed-by: Stephen White <senorblanco@chromium.org>