Release ShaderModuleBase::mTintProgram when it is not needed
This CL will release mTintProgram and mTintSource in
a ShaderModuleBase, when the ShaderModule is not referenced
externally, and the ShaderModule is not needed for compiling any
pipeline. And mTintProgram and mTintSource will be re-created
from the shader source code when the ShaderModule is need for
initializing a new pipeline.

Bug: dawn:2184
Change-Id: Ice3a048a1ea94c02cb56f03574379295941202cd
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/169943
Commit-Queue: Peng Huang <penghuang@chromium.org>
Kokoro: Kokoro <noreply+kokoro@google.com>
Reviewed-by: Corentin Wallez <cwallez@chromium.org>