Remove data returned from `TextureBuiltinFromUniform` transform.
This CL updates the tint/dawn API to remove the data returned from the
`TextureBuiltinFromUniform` transform. A new API is added to the
inspector which Dawn can use to query the same information. Dawn then
provides tint with a list of binding points in the order to place them
in the generated UBO.

Change-Id: I7049ff07f0d55766206f73f7c07ad2df383f5c2d
Bug: dawn:1299
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/168900
Kokoro: Kokoro <noreply+kokoro@google.com>
Commit-Queue: dan sinclair <dsinclair@chromium.org>
Reviewed-by: Stephen White <senorblanco@chromium.org>
Reviewed-by: Shrek Shao <shrekshao@google.com>
Reviewed-by: Ben Clayton <bclayton@google.com>