[Mac] Create and hold MtlTextures in SharedTextureMemoryMTL
Per validation constraints [1], the MtlTextures that TextureMTL creates
when initialized from SharedTextureMemory must match the properties of
the backing IOSurface that the SharedTextureMemoryMTL instance was
created with. This means that we can create the MtlTextures in
SharedTextureMemoryMTL on its creation and simply vend them out to
any TextureMTL objects that that SharedTextureMemoryMTL instance is
asked to create. This CL makes that change.

[1] https://source.chromium.org/chromium/chromium/src/+/main:third_party/dawn/src/dawn/native/SharedTextureMemory.cpp;l=182-193?q=SharedTextureMemory.cpp&ss=chromium

Change-Id: I5fa35b1e55b16436674dd3f730e613e6fcd9be9c
Bug: dawn:2152, 1493854
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/171041
Commit-Queue: Colin Blundell <blundell@chromium.org>
Kokoro: Kokoro <noreply+kokoro@google.com>
Reviewed-by: Corentin Wallez <cwallez@chromium.org>