Implement push_constant-based firstIndex transform.
This transform offsets the vertex_index and/or instance_index builtin
variables by a user-supplied first_vertex and first_instance offset.
This is required for shading languages that use zero-based first_*
variables such as GLSL and HLSL.

Also, GLSL requires push constants to be wrapped in a struct, and the
struct to be emitted, but not emitted as an interface block. Thus,
change the "skip_push_constants" flag in the AddBlockAttribute
to "push_constants_wrap_only", which causes push constants to be
wrapped in a struct, but to not have the BlockAttribute added.

Bug: tint:2140
Change-Id: I7ce7bb67b58efa110e91fc38d840e674adaf9d0b
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/169461
Kokoro: Kokoro <noreply+kokoro@google.com>
Commit-Queue: Stephen White <senorblanco@chromium.org>
Reviewed-by: Austin Eng <enga@chromium.org>
Reviewed-by: Ben Clayton <bclayton@google.com>