cmake/dxc: only include third_party/dxheaders on non-Windows
The DirectXHeaders repo is used by DXC for Linux/MacOS builds, and is
pinned at a very specific commit. For Windows, we need to use the
official platform SDK, which is picked up when this variable is not
already set.

Change-Id: Ide7220240991480b3c540c1cf7774fb684eac5a5
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/171520
Kokoro: Kokoro <noreply+kokoro@google.com>
Reviewed-by: Ben Clayton <bclayton@google.com>
Reviewed-by: Austin Eng <enga@chromium.org>
Commit-Queue: Antonio Maiorano <amaiorano@google.com>