Remove condition in absl_config for cflags
This CL replaces the condition `dawn_standalone` with
`!build_with_chromium` when setting cflags for the absl_source_set
build configs. This removal is necessary for ANGLE to be able to
build dawn native as it needs the "-Wno-unreachable-code-break" flag
to be set for dawn's abseil build.

Bug: angleproject:8477
Change-Id: I644d216339da60eccb34a0467b863b60e06efb0c
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/174080
Commit-Queue: Liza Burakova <liza@chromium.org>
Kokoro: Kokoro <noreply+kokoro@google.com>
Reviewed-by: dan sinclair <dsinclair@chromium.org>