[toggles] Set labels on backend objects only if toggle is set
We have found that in Chrome, setting these labels can be a source of
non-trivial overhead on performance benchmarks. This CL guards setting
of these labels with the `use_user_defined_labels_in_backend` toggle.
Note that labels will still be set on frontend objects, and hence will
still be available in e.g. Dawn error messages. The labels on the
backend objects are relevant only when debugging via platform-specific
facilities (e.g., XCode).

This CL also changes the `use_user_defined_labels_in_backend` toggle to
default to true if validation is enabled. This is necessary as at least
on Vulkan Dawn's debugging facilities assume that these labels are
set [1]. It is consistent in any case, as a client enabling validation
indicates that they are more concerned about debuggability than
performance.

[1] https://source.chromium.org/chromium/chromium/src/+/main:third_party/dawn/src/dawn/native/vulkan/BackendVk.cpp;l=263?q=BackendVk.cpp&ss=chromium

NOTE: This CL is a reland following the suppression of a pre-existing
failure that was tickled by this CL (a Vulkan backend error message was
previously swallowed in Dawn backend code, but the lack of a debug label
for devices meant that the error was now surfaced). The reland is
unchanged save for incorporating
https://dawn-review.googlesource.com/c/dawn/+/173820.

Bug: dawn:2383
Change-Id: I8196c6ec165add0d5e9e73952be77df20e1c2dc3
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/174280
Kokoro: Kokoro <noreply+kokoro@google.com>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Reviewed-by: Corentin Wallez <cwallez@chromium.org>