OpenGL: scale the depthBias value by half.
There is an ambiguity in the GL and Vulkan specs with respect to
depthBias: If a depth value lies between 2^n and 2^(n+1), is the
"exponent of the depth value" in the spec n or n+1? It looks like GL
drivers use n, while the WebGPU CTS is expecting n+1. Scaling the depth
bias value by 0.5 gives results that the WebGPU CTS expects.

Bug: dawn:2076
Change-Id: If2965acd5250b1ed834a0bad260907522c179679
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/174922
Reviewed-by: Austin Eng <enga@chromium.org>
Commit-Queue: Stephen White <senorblanco@chromium.org>
Kokoro: Kokoro <noreply+kokoro@google.com>