Fix issue with break inside switch inside continuing.
Tint is currently too restrictive in testing for a break statement
inside of a continuing block. In the case where the break is inside a
switch it is permissible, as it breaks the switch, not the loop.

This CL updates the validation routines to allow a switch to hold a
break statement inside a continuing block.

Bug: tint:2144
Change-Id: I82850cb736d1c2433fd98f2a2cd9bfd325d698d7
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/175080
Commit-Queue: dan sinclair <dsinclair@chromium.org>
Reviewed-by: Ben Clayton <bclayton@google.com>
Kokoro: Kokoro <noreply+kokoro@google.com>