D3D12: Don't skip flushes when the recording context is not open.
Make the CommandRecordingContext always open instead by initializing it
on Queue creation and reinitializing it after a submit. This allowed
removing all the IsOpen() logic. In order to move in the queue the logic to
reinitialize the context, the CommandAllocatorManager was folded in the
Queue, where allocation of ID3D12CommandLists and their reuse can be
controlled more precisely.

Bug: dawn:1413
Change-Id: If9b7d49295451b630eb999b20066151f97ccbc10
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/175222
Commit-Queue: Corentin Wallez <cwallez@chromium.org>
Kokoro: Kokoro <noreply+kokoro@google.com>
Reviewed-by: Austin Eng <enga@chromium.org>
Reviewed-by: Loko Kung <lokokung@google.com>