Compat: Handle index buffer offset with indexed indirect draw.
The validation is already copying and updating all of the
indirect draw buffers so this just adds the adjustment to
update the firstIndex parameter when copying.

Validation would always be on for WebGPU in the browser.
Emits an error if validation is off and the index buffer
offset is non-zero.

Bug: dawn:161
Change-Id: I385f5a8907edd85a9c44652d495b11dd4710b5a8
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/174400
Reviewed-by: Austin Eng <enga@chromium.org>
Kokoro: Kokoro <noreply+kokoro@google.com>
Reviewed-by: Stephen White <senorblanco@chromium.org>
Commit-Queue: Gregg Tavares <gman@chromium.org>