Remove ShaderModule IsAlive assertion
TSAN says this is a data race. Checking liveness in the linked-list
would need to acquire the linked-list lock. In this code, this
assertion doesn't have a lot of value anyway, so remove it.

This is failing on top-of-tree, but the tests that would have caught
it only run on NVIDIA, and the particular test is skipped because
TSAN reports a race within the NVIDIA driver.

Bug: dawn:1662
Change-Id: Ibf319888fe23f2194b3cc75be49387a32a4a9759
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/175205
Kokoro: Kokoro <noreply+kokoro@google.com>
Commit-Queue: Austin Eng <enga@chromium.org>
Reviewed-by: Corentin Wallez <cwallez@chromium.org>
Reviewed-by: Loko Kung <lokokung@google.com>