[tools][cts] Differentiate between 'no results' and 'no tests' that match an expectation
The former will be preserved, as it may be for an experimental variant. The latter is determined by looking at the CTS test list, and will be removed, as this is an expectation validation error.

Change-Id: I4a85e42fff7a50c6850efd39afcf9421278d82ac
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/177801
Kokoro: Kokoro <noreply+kokoro@google.com>
Auto-Submit: Ben Clayton <bclayton@google.com>
Commit-Queue: Austin Eng <enga@chromium.org>
Reviewed-by: Austin Eng <enga@chromium.org>