[cts] Skip SwiftShader copyToTexture,ImageBitmap:from_ImageData tests
Something with the most recent roll has caused more test cases to start crashing with SwiftShader.
Given that the ImageBitmap tests haven't changed in months, this is likely to be related to the harness changes for re-acquiring a device.

I'd normally investigate exactly *why* this is broken, but we already have [ Skip ] for the sibling tests, and this does not affect anything other than SwiftShader - so it seems most likely that a change to the CTS has just uncovered more SwiftShader brokenness.

Bug: chromium:1344876
Change-Id: I65c71cfa3574cba99474250e7cbe7e6c7da2e5ee
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/178640
Auto-Submit: Ben Clayton <bclayton@google.com>
Kokoro: Kokoro <noreply+kokoro@google.com>
Commit-Queue: Ben Clayton <bclayton@google.com>
Reviewed-by: Corentin Wallez <cwallez@chromium.org>