Remove SwapChain from the API
This applies the switch from manipulating a SwapChain object to calling
Surface::configure that was brung to the standard webgpu.h a couple of
months ago.

I believe this needs a bit of extra work, but at this stage I need a
review from people more familiar with the codebase to tell me if I am
going in the right direction.

The SwapChain object is still here, but no longer present in the API.
Instances of the swap chain are now owned by the surface object. The
test for surface capabilities is partly calling a method of the
PhysicalDevice in order to have backend-specific capabilities.

So far I mostly tested with the Vulkan backend, though I tried to
provide a valid implementation for all backends.

Bug: dawn:2320
Change-Id: I69d3c64873213eca375c66170cb648debd8ec91d
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/167600
Kokoro: Kokoro <noreply+kokoro@google.com>
Commit-Queue: Élie Michel <elie.michel.fr@gmail.com>
Reviewed-by: Corentin Wallez <cwallez@chromium.org>
Reviewed-by: Austin Eng <enga@chromium.org>