Fix dangling pointer in ProgrammableEncoder
This patch fixes the dangling pointer (`mEncodingContext`) in
`ProgrammableEncoder`. As `mEncodingContext` always points to a
piece of memory owned by a member of the inherited class of
`ProgrammableEncoder` (`Ref<CommandEncoder>` in `RenderPassEncoder`
and `ComputePassEncoder`, and `mBundleEncodingContext` in
`RenderBundleEncoder`), we must always clear `mEncodingContext`
before deleting the piece of memory it points to.

Bug: dawn:2349
Change-Id: Iad15cf5fbf14818bbfad7e701585d4520b3e7d03
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/178981
Commit-Queue: Jiawei Shao <jiawei.shao@intel.com>
Reviewed-by: Austin Eng <enga@chromium.org>
Kokoro: Kokoro <noreply+kokoro@google.com>
Reviewed-by: Corentin Wallez <cwallez@chromium.org>