Don't expose chromium-experimental-subgroups if DXC is not used
Currently, the feature gets exposed if DXC is available. However,
there are some cases where DXC is still not used (if the shader model
is not high enough to use it).

DXC may also be disabled manually via command line or toggles, so
subgroups should not be exposed in this case.

Bug: chromium:41494633
Change-Id: I8cca2263e4a8ae0d1e6472901efd68784ffe06b1
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/179440
Reviewed-by: Antonio Maiorano <amaiorano@google.com>
Commit-Queue: Austin Eng <enga@chromium.org>
Kokoro: Kokoro <noreply+kokoro@google.com>