[tint][ir] Skip unreachable continuing blocks
When converting AST to IR, skip over unreachable continuing blocks as
they may contain references to unreachable declarations that were also
skipped.

Note that this issue is also present in the legacy SPIR-V writer, and
is not fixed in this CL, hence the SKIP expectations for that path.

Fixed: tint:2201, tint:2202
Change-Id: I2689d412e340508b005a33e251b669948d986cd5
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/179240
Reviewed-by: dan sinclair <dsinclair@chromium.org>
Kokoro: Kokoro <noreply+kokoro@google.com>
Reviewed-by: Ben Clayton <bclayton@google.com>