Fix issue with external texture mapping in MSL.
This CL fixes an issue with the external textures for MSL. Because of
the way MSL works where there is no `group` everything maps to `0`. An
internal flag is set on new variables to allow collisions (they don't
collide within a type of MSL binding, but may in WGSL). The flag was not
previously set on the originating variable.

This CL updates the code to set the originating variable as also
potentially colliding.

Bug: dawn:2472
Change-Id: I2aad439fdef0b697caa6663d0a5705d4e8b0dd48
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/180400
Commit-Queue: dan sinclair <dsinclair@chromium.org>
Auto-Submit: dan sinclair <dsinclair@chromium.org>
Commit-Queue: Ben Clayton <bclayton@google.com>
Reviewed-by: Ben Clayton <bclayton@google.com>