[Static samplers] Pivot to having StaticSamplerBindingLayout
This CL moves static samplers away from being specified via a chained
struct on SamplerLayoutBinding to a chained struct on
BindGroupLayoutEntry, similarly to ExternalTextureBindingLayout. This
organization avoids redundantly specifying the sampler type and makes
the conceptual relationships more clear: a SamplerLayoutBinding is for
a sampler supplied dynamically, whereas the StaticSamplerBindingLayout
is for samplers supplied statically.

As a bonus, this new code organization directly points the way to a
couple next steps that need to happen:
* Incorporation of validation of the number of static samplers
* Populating BindingInfo with the static sampler's sampler object

This CL leaves TODOs for those next steps.

Bug: dawn:2463
Change-Id: I9974b5678051ea38ebab57e801affc474d13938b
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/181042
Reviewed-by: Corentin Wallez <cwallez@chromium.org>
Reviewed-by: Austin Eng <enga@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>