Fix dangling pointers in DeviceBase
This patch fixes all the dangling pointers in DeviceBase. In
Chromium the callbacks and userdata in DeviceBase should be cleared
after Device.Destroy() is called and before Dawn wire server is
destroyed, so we should support setting nullptr to callbacks and
userdata even when Device is lost.

Bug: dawn:2349
Change-Id: I5f4e919e44b66181b4fac003ef25e088461d26cc
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/181144
Reviewed-by: Austin Eng <enga@chromium.org>
Reviewed-by: Corentin Wallez <cwallez@chromium.org>
Commit-Queue: Jiawei Shao <jiawei.shao@intel.com>