[tint][ir] Use the FilteredIterator to avoid dead instructions
The common pattern by many IR transforms is to iterate over the ir::Module::instructions.Objects(), which is the list of allocated instructions.
These include instructions that have been destroyed, and must not be used.

We've hit the same bug time and time again, where we've forgotten to check for `Instruction::Alive()` before using the instruction.
To avoid this mistake, add new ir::Module::Instructions() and ir::Module::Values() method that returns an iterator filtered to the alive instructions / values.

Move the allocators into a new 'allocators' structure member to discourage use of these for object iteration.

Change-Id: Ic0043c8a4b82b9ec10b2b8a3479f33c6e5d286ff
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/182122
Commit-Queue: Ben Clayton <bclayton@google.com>
Reviewed-by: James Price <jrprice@google.com>