Skip SwapChainTests on experimental Linux/Intel
Skips all SwapChainTests on Linux + Intel + Vulkan + Mesa 23.2.1. This
is necessary because certain tests are crashing on this config, but the
crash originates in the SetUp code, and thus individual tests cannot be
skipped.

In order to limit the scope of these skips, also adds a new IsMesa check
to DawnTestBase.

Bug: dawn:2492
Change-Id: I6d69b07ebcea1576c87fc0b9a4b0ee4027d9d13d
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/183020
Auto-Submit: Brian Sheedy <bsheedy@google.com>
Commit-Queue: Austin Eng <enga@chromium.org>
Reviewed-by: Austin Eng <enga@chromium.org>