Suppress UBSAN false positive
Similar to the previous CFI suppression, this cast is valid because
we cast to type T before T is constructed, but we never actually
use T in this cast.

It should be valid because use of the cast pointer only occurs
after construction is complete.

Fix failures in
https://ci.chromium.org/ui/p/chromium/builders/ci/linux-ubsan-vptr/29687/overview

Bug: chromium:40195122
Change-Id: I7fe1d16fe0a3fb3f630a9f65af2f2096e2fd562c
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/185964
Reviewed-by: Corentin Wallez <cwallez@chromium.org>
Commit-Queue: Austin Eng <enga@chromium.org>