Add c++ "init structs" for chained struct types
Chained structs can't currently use designated initializers. This
is because aggregate initialization is not supported when the
struct has a constructor. The constructor defaults the sType.

To allow for a sort of aggregate initialization with designated
initializers, add an Init struct from which the struct can be
constructed. This allows for aggregate-looking construction
with mostly the same syntax. You can now construct a chained struct
with double-braces (instead of single) like:

```
MyChainedStruct {{
  .foo = 42,
  .bar = "hello!",
}};
```

Bug: chromium:40195122
Change-Id: I05e99335f48dd56ee7cdc0c4539111c630c70d8a
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/185967
Reviewed-by: Loko Kung <lokokung@google.com>
Reviewed-by: Kai Ninomiya <kainino@chromium.org>
Commit-Queue: Austin Eng <enga@chromium.org>