[tint][ir] Refactor the Disassembler, rename to Disassembly
Make the bulk of the disassembler private, and invoke the disassemble
function from the constructor. As the class only disassebles once, and
the API is (mostly) immutable, rename the class to Disassembly.

This CL contains a fix for block-parameter values being printed with
their type. No other value did this.

Change-Id: I8df891a06bc0a70d75538340c89d9d1e57757121
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/186461
Reviewed-by: James Price <jrprice@google.com>