[ir] Add Instruction::DetachResult() helper
This removes an instruction result from an instruction and returns
it. This makes it easier to move an instruction result to a new
instruction, which should be more efficient than using
`ReplaceAllUsesWith()` to update usages one at a time.

Add overloads of a few IR Builder functions that take an
InstructionResult so that the IR builder can be used with this
pattern, instead of manually creating instructions with the allocator.

Update many transforms to use this pattern.

Change-Id: I19088816c0e715b59ea84ed1bdb092fdd2567fa6
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/186565
Commit-Queue: James Price <jrprice@google.com>
Reviewed-by: Ben Clayton <bclayton@google.com>
Reviewed-by: dan sinclair <dsinclair@chromium.org>