[ir] Add ReferencedModuleVars helper
This is a stateful helper that returns (and caches) the module-scope
variables that are transitively referenced by a function. Some work is
done on creation to determine which variables are directly referenced
by which blocks. The remaining work is done when the references for a
function are actually requested.

Update the ZeroInitializeWorkgroupMemory transform to use this helper.

Change-Id: I522d6521a61a94bf255ca9d9ff3d7d5998a2b69e
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/169222
Reviewed-by: dan sinclair <dsinclair@chromium.org>
Reviewed-by: Ben Clayton <bclayton@google.com>
Commit-Queue: James Price <jrprice@google.com>