[tint][ast] Fix DirectVariableAccess with uncalled function ptr params
If a function has no calls, and in turn calls another function with a function-address-space pointer parameter, then we could incorrectly attempt to create a parameter with a pointer-to-pointer type.
The code was missing an `UnwrapPtrOrRef` - this wasn't noticed as most originating variables are `ref` types, and `CreateASTTypeFor()` will construct the storage type for a `ref` type.

Bug: 339684346
Change-Id: I0c32536acb6f582ae521868e1599c22c77f656bc
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/187689
Reviewed-by: James Price <jrprice@google.com>
Commit-Queue: James Price <jrprice@google.com>
Auto-Submit: Ben Clayton <bclayton@google.com>