[Vulkan] Populate YCbCr info in SharedTextureMemory::GetProperties()
This CL populates YCbCr info when it is passed into
SharedTextureMemory::GetProperties() on Android/Vulkan. The populated
information comes from the AHB backing the SharedTextureMemory. We also
add a test verifying that the populated info matches that of the AHB
backing the STM instance, as well as validation that the passed-in
YCbCr info does not have anything chained onto it.

Bug: dawn:2476
Change-Id: Ie9536d42d59bc4e1a49d4bc3b4cf98464c89bd33
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/188240
Reviewed-by: Austin Eng <enga@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>