[Vulkan] Have YCbCrVkDescriptor::vkChromaFilter be FilterMode
Skia will need to query the value of Dawn's
YCbCrVkDescriptor::vkChromaFilter to determine whether linear
filtering is supported. Currently this field is a uint32_t that is set
to the raw Vulkan values of the VkFilter enum, meaning that Skia would
need to query against that enum :\. This CL changes vkChromaFilter to
be a wgpu::FilterMode instead.

Change-Id: I63bbb67d2dbc95da5f3410373fe27440372f6234
Bug: dawn:2476
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/188760
Reviewed-by: Corentin Wallez <cwallez@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>