Add generating Tint IR fuzzer corpus
This adds in a Python script for generating IR binary file from a WGSL
file, build rules for running this over the testing WGSL files, and
changes to the IR binary outputting code to reduce the number of
failures due to unsupported inputs.

Currently 7203 out of 7275 shaders that are used in the WGSL corpus
are convertable. Those that are not, either use enables that are not
supported in the IR, or fail for reasons like missing override
values. The generation code is resiliant enough that these failure do
not break process, so investigating those failing shaders can be done
in a follow up CL/issue.

Issue: 341114373

Change-Id: Ic7b26010465865c46b1013db3a0613d89e0db457
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/188662
Auto-Submit: Ryan Harrison <rharrison@chromium.org>
Reviewed-by: Ben Clayton <bclayton@google.com>
Commit-Queue: Ryan Harrison <rharrison@chromium.org>
Reviewed-by: James Price <jrprice@google.com>