Make textureDimensions(texture_external) use the visible size.
Previously it was using the size of plane0, which was incorrect when the
plane was cropped or for some rotations. Adds a test that checks that
the correct size is returned.

Also fixes the computation of the visibleSize in ExternalTexture to
account for the rotation.

Fixed: 341216488
Change-Id: If81ebb83f044795b641e888491f2f8603b38be0a
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/188802
Reviewed-by: Ben Clayton <bclayton@google.com>
Reviewed-by: James Price <jrprice@google.com>
Commit-Queue: Corentin Wallez <cwallez@chromium.org>