[tint][ir] Use Operand() instead of indexing operands_ directly
Operand() will return nullptr if the index is OOBs. This can be used by the validator and disassembler to highlight unexpected null operands.

Given that we assert the index is in bounds, the compiler should be able to elide the assert and end up with something very similar in terms of performance.

Change-Id: Iba408ae7a3cf37f4e8711d37ddffa3f88ef4b649
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/189709
Reviewed-by: James Price <jrprice@google.com>
Commit-Queue: Ryan Harrison <rharrison@chromium.org>