[Android] Add Device::GetAHardwareBufferProperties()
This CL adds Device::GetAHardwareBufferProperties(), which will allow
getting the properties of an AHardwareBuffer (notably the YCbCr info)
without having to first create a SharedTextureMemory instance. This is
useful because it turns out that when using external sampling, it will
be necessary to configure the SharedTextureMemory at the time of its
creation so that it correctly creates its VkImage. Having this API on
Device allows clients to determine whether they want to use external
sampling with a given AHB before creating the STM for that AHB.

Note:

* I didn't bother deduping the implementation with that of
  SharedTextureMemory as once this new API is in place Chromium will
  transition to using it entirely and we can remove obtaining the AHB
  properties from SharedTextureMemory

Change-Id: Iaa0b6ad166bb6b7c668acce2336863dd021fe8a1
Bug: dawn:2476
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/190140
Reviewed-by: Austin Eng <enga@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>