[YCbCr Samplers] Add SampleTypeBit::External
This adds SampleTypeBit::External and adds support in
AddConditionalColorFormat.

Note: It does not add corresponding wgpu::TextureSampleType as we
only need it internally and does not need to be exposed. Accordingly
it is set to a high value for future SampleTypeBit that can be added.

Change-Id: I68fef904d3963910e96146a973d333d87365dab5
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/190160
Commit-Queue: Saifuddin Hitawala <hitawala@chromium.org>
Reviewed-by: Austin Eng <enga@chromium.org>