[tint][resolver] Mark all short-circuited RHS expressions as not-evaluated
Not just those that would otherwise be const-expression.

The rules around this were patchy, and there were situations where if
the RHS was a runtime-expression builtin call, the binary op would be
classed as EvaluationStage::kRuntime, the LHS expression
EvaluationStage::kConstant, and the RHS classed as EvaluationStage::kRuntime, but the RHS sub-expressions were marked EvaluationStage::kNotEvaluated.

This would lead to bad assumptions downstream.

With this change, a short-circuited binary expression with a
const-expression LHS will always become EvaluationStage::kConstant, with the LHS being EvaluationStage::kConstant and the RHS always being EvaluationStage::kNotEvaluated.

Fixed: 341124493
Change-Id: I47d4b7e3ce1a74f74db4a51e816f114aa3bec015
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/190400
Commit-Queue: Ben Clayton <bclayton@google.com>
Reviewed-by: James Price <jrprice@google.com>