[spirv] Fix terminator creation in MergeReturn
When we walk up the control stack to add missing terminators to `if`
instructions, we need to limit this to `if` instructions that were
added by the transform. Otherwise, we may add a terminator to a block
that is still being processed, and end up with multiple terminators.

Fixed: 343597426
Change-Id: If6c345e817a11e68396f5eeed29d38c5b7478323
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/190723
Commit-Queue: James Price <jrprice@google.com>
Reviewed-by: dan sinclair <dsinclair@chromium.org>