Convert std::endl to newlines.
Currently the codebase sometimes uses `\n` and sometimes uses
`std::endl`. This CL converts the remaining `std::endl` usages to just
use the `\n` form directly. This removes the need to go through the
`std::endl` and removes the calls to `flush` which it adds.

(This is in line with the cppcoreguidlines SL.io.50)

Change-Id: Icc8b3c99e9a96c8b6fd504ce47e80d7f307b7633
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/191020
Commit-Queue: dan sinclair <dsinclair@chromium.org>
Reviewed-by: Corentin Wallez <cwallez@chromium.org>
Reviewed-by: James Price <jrprice@google.com>