d3d: Change D3D references to PhysicalDevice to WeakRef instead of Ref
BackendD3D holds strong references to PhysicalDeviceD3D11/12, which
are created when enumerating adapters. Once adapters are enumerated,
these are kept alive indefinitely. Upon creation, PhysicalDeviceD3D
has a reference to the D3D11/12Device, which means that the
corresponding physical adapter is kept powered on indefinitely.

This CL changes the strong Ref to a WeakRef and relies on the caller
to keep a strong reference to the PhysicalDevice to keep it alive.
All other unused devices are cleaned up, which releases the unused
D3DDevices.

Bug: 342299153
Change-Id: I4ff6979abb175f9b737fb3ede4b26334d858d6a4
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/189581
Reviewed-by: Austin Eng <enga@chromium.org>
Commit-Queue: Patrick To <patrto@microsoft.com>