Return wgpu::Status from SharedTexture/BufferMemory Begin/EndAccess
Note, there is a small behavior change where previously, you could
EndAccess and return "true" if the access was successfully ended, but
there was some other error exporting the fence. We could return a new
enum to differentiate these cases, but ultimately I decided it is not
worth it. If the fence failed to export, then it would be unsafe to
use the texture after anyway since you can't wait on it. All the error
cases are merged into one now.

Bug: 42241199
Change-Id: I19f2700bd00b9661bd37943072b9d149b956f55f
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/193282
Reviewed-by: Kai Ninomiya <kainino@chromium.org>
Commit-Queue: Austin Eng <enga@chromium.org>
Reviewed-by: Loko Kung <lokokung@google.com>