[ir] Strip all IO attributes in ShaderIO transform
We were only stripping shader IO attributes from members of structures
that were used for the shader interface, which was leading to ICEs in
printers when structures that were not used for the interface had
these attributes. Instead, strip attributes from all structures that
existed prior to the transform.

Refactor the State object to behave the same as other transform state
objects, with a single `Process()` function for the whole transform.

Bug: 42251016
Change-Id: Ie140ea2c317590f54665a394323ed137abd9a862
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/193322
Commit-Queue: James Price <jrprice@google.com>
Reviewed-by: dan sinclair <dsinclair@chromium.org>