Make a convenience cpp method for WaitAny with one future
When one future is waited on, there is no need for FutureWaitInfo.
FutureWaitInfo tells you which futures completed when the application
passed a list of futures. But when the list is only one element,
WaitStatus::Success always means that the future completed.

When waiting on just one future, this is much more convenient than
creating the extra struct.

Bug: none
Change-Id: I75b722511a8025a5c1d9229380a5a1618d21a0a8
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/193701
Commit-Queue: Austin Eng <enga@chromium.org>
Reviewed-by: Loko Kung <lokokung@google.com>
Reviewed-by: Kai Ninomiya <kainino@chromium.org>