[hlsl] Add option to specify the downstream HLSL compiler.
With the new HLSL IR backend we plan on moving the FXC specific work
arounds into an FXC transform. That transform will only be run when
we're targeting FXC. In order to do so, a `Compiler` option is added to
the HLSL backend options and set as needed by Dawn.

Bug: 42251045
Change-Id: I42b28ddb9146fbdde5cd7cecabe7b8a242f51653
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/194420
Reviewed-by: James Price <jrprice@google.com>
Commit-Queue: dan sinclair <dsinclair@chromium.org>