Require both `blend_src(0)` and `blend_src(1)` must be used in one struct
This patch adds a new check on dual source blending that both
`@location(0) @blend_src(0)` and `@location(0) @blend_src(1)` must be
used in one struct, and other members (if exist) must not have
`@location` attribute, or neither of them are used.

As now this rule is applied to all structs, the original checks on the
entry point are all removed.

Bug: chromium:341973423
Test: tint_unittests
Change-Id: I71a7b088e57bfd67eb1adf9f1c053f1592c3ce5a
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/191766
Commit-Queue: Jiawei Shao <jiawei.shao@intel.com>
Reviewed-by: James Price <jrprice@google.com>