Indenting fixes in Kotlin template
This is in preparation for CL "Generalize JNI type logic in macros"
which for some reason triggers assertions in generator_lib.py
preprocess caused by *existing* indenting issues in the template.

generator_lib wants all content in a nested jinja2 block to have a
greater indention, which I can work around without an extra indention
for the C block by closing the scope on the same line as the next {%
elif %} etc.

It doesn't recognise {% if .. %} etc if there's a newline inside the
tag.

Change-Id: Id3a8eb3aeb29dbac9e562df285dd1a1331eee0a6
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/195294
Commit-Queue: Jim Blackler <jimblackler@google.com>
Reviewed-by: Corentin Wallez <cwallez@chromium.org>