d3d: Use EnumAdapterByGpuPreference when available
Dawn currently uses EnumAdapters1 to enumerate IDXGIAdapters, and then
uses various properties of the adapter to infer the power preference.
This change uses the IDXGIFactory6::EnumAdapterByGpuPreference to
correctly return the most accurate power preference order, in
IDXGIFactory6 is available. EnumAdapters1 is used as a fallback. The
order of adapters returned by either API should take priority, so the
manual sorting is Dawn is bypassed for D3D backends.

Bug: 342299155

Change-Id: I5e27fbe6bb301cb393aafbc9468015818407044c
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/189641
Commit-Queue: Patrick To <patrto@microsoft.com>
Reviewed-by: Austin Eng <enga@chromium.org>
Reviewed-by: Rafael Cintron <rafael.cintron@microsoft.com>