[Tint] Only allow read textures in vertex shader
For builtin functions that accept textures as arguments, only
non-writable texture types should be allowed in vertex shaders. This is
now addressed in the WGSL definitions and the appropriate generated
tests are fixed.

The manually modified files in this change are:
* src/tint/lang/wgsl/wgsl.def
* src/dawn/tests/unittests/validation/StorageTextureValidationTests.cpp
* src/tint/lang/wgsl/intrinsic/table_test.cc
* src/tint/lang/wgsl/resolver/builtin_test.cc

This does result in a slightly confusing error message of:

`error: built-in cannot be used by vertex pipeline stage`

when a built-in function with a writable texture arg is used (even if the built-in function can be used with a read-only texture), however it will be improved once https://dawn-review.googlesource.com/c/dawn/+/190781/ is merged to:

`error: storage texture with read_write access mode cannot be used by vertex pipeline stage`

Bug: 344846829
Change-Id: I47f544fbea715aa8825c3d13502d9e8740074c95
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/195417
Commit-Queue: Natalie Chouinard <chouinard@google.com>
Reviewed-by: James Price <jrprice@google.com>