[tint][ir][fuzz] Convert assembler to generating test case protos
The original prototype for the IR fuzzer had the corpus being composed
of IR protos directly, but the fuzzer has since been changed to take
in a test case specific proto that is an IR proto with a sidecar
binary blob. This updates the assembler to generate this new proto.

The assembler has also been changed to not default print to STDOUT,
since it just spits out binary, which kept on breaking my terminal.

Fixes: 345377541

Change-Id: Icae5795d88a3157154be776fe6e6d94661c978aa
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/195936
Commit-Queue: James Price <jrprice@google.com>
Auto-Submit: Ryan Harrison <rharrison@chromium.org>
Reviewed-by: James Price <jrprice@google.com>