Fix std::vector<const T>
It's not allowed by the standard, and libc++ is dropping support for it.

In this case, MTLCommonCounter is declared as:

  typedef NSString *const MTLCommonCounter;

which is unfortunate as the 'const' doesn't really do much, but it does
mean the type can't be used in the standard collections.

Bug: chromium:323708866, chromium:349411322
Change-Id: I953d1e0392b58480ffdb1482943f68726c2c67b1
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/196054
Commit-Queue: Hans Wennborg <hwennborg@google.com>
Auto-Submit: Hans Wennborg <hwennborg@google.com>
Commit-Queue: Corentin Wallez <cwallez@chromium.org>
Reviewed-by: Corentin Wallez <cwallez@chromium.org>