GL: Add support for wgpu::Surface.
This adds a dawn::native::opengl::SwapChainEGL object that uses
EGLSurfaces under the hood. Due to how special the default framebuffer
is in OpenGL (you can draw Y-flipped to it as the sole attachment or
with depth-stencil, or blit to it), rendering is done "offscreen" and
blitted with Y-flip on Present().

ContextEGL gains a "MakeSurfaceCurrentScope" method that helps make sure
a surface is always present even as opengl::Device::GetGL/EGL keeps
doing calls to MakeCurrent.

DisplayEGL gains a ChooseConfig method to help choose EGL configs for
the swapchain (and later will be used for context creation).

Many improvements are possible to this implementation:

 - Supporting rendering directly to the default framebuffer when it is
   RenderAttachment only and the sole attachment of render passes. (will
   require adding a Y-flip uniform to every vertex shader).
 - Supporting rendering directly to the default framebuffer when used
   with a depth-stencil attachment that's discarded as EGLConfig can
   contain depth-stencil.
 - Support for SRGB swapchains with EGL_KHR_gl_colorspace.
 - Support for more than the hardcoded RGBA8Unorm formats with
   EGL_KHR_no_config_context.
 - Dynamic support for wayland and x11 with EGL_EXT_platform_base and
   friends.

Bug: 344814083
Change-Id: I64f3bb5d314acea87f472eaa78dff06deeb4ff3d
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/191142
Reviewed-by: Stephen White <senorblanco@chromium.org>
Commit-Queue: Corentin Wallez <cwallez@chromium.org>