Implement BufferMapExtendedUsages on D3D11
The new MappableBuffer supports mapping usage and any other usages.
This is achieved by managing several copies of the buffer, each with its
own ID3D11Buffer storage for specific usage.
For example, a buffer that has MapWrite + Storage usage will have at
least two copies:
- One copy with D3D11_USAGE_DYNAMIC for mapping on CPU.
- One copy with D3D11_USAGE_DEFAULT for writing on GPU.

Internally this class will synchronize the content between the copies so
that when it is mapped or used by GPU, the appropriate copy will have
the up-to-date content. The synchronizations are done in a way that
minimizes CPU stall as much as possible.

Bug: 345471009
Change-Id: I78c51a2bed5adcbdb65d05b5fc0f41706a53d9d7
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/191980
Commit-Queue: Quyen Le <lehoangquyen@chromium.org>
Reviewed-by: Corentin Wallez <cwallez@chromium.org>
Reviewed-by: Austin Eng <enga@chromium.org>