[tools][cts] Default to immutable chunks
Switches the CTS roller's logic to assume all expectation chunks are
immutable unless one of the following is true:
1. The chunk is one of the auto-generated ones from the roller
2. The chunk is explicitly marked as MUTABLE

This is the opposite of the current behavior, where all chunks are
assumed to be mutable unless they are marked with KEEP. In practice,
this seems to make the roller behave more consistently, as it only
needs to produce new expectations instead of trying to refactor
existing ones.

This can potentially lead to cases where the roller will get stuck
such as if a RetryOnFailure expectation failure exists already and
the test starts to fail consistently. However, this should (hopefully)
still result in more consistent rolls since this is not a common
scenario. The resulting fix should also be simpler than trying to
resolve conflicts, as it will just require updating the existing
RetryOnFailure expectations to Failure.

The old KEEP annotations will be kept around for now in case this
CL ends up needing to be reverted, but they can be removed in the
future.

Bug: chromium:345280734
Change-Id: I531422dfc8f500ecab1b525a166a0abf72a8a08f
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/196236
Commit-Queue: dan sinclair <dsinclair@chromium.org>
Auto-Submit: Brian Sheedy <bsheedy@google.com>
Reviewed-by: dan sinclair <dsinclair@google.com>
Reviewed-by: dan sinclair <dsinclair@chromium.org>