[MSan] Fix use-after-dtor issues in generated C++ APIs.
The move assignment operator manually invokes the destructor, to ensure
any data owned by pointer fields is cleaned up instead of leaked.
However, manually invoking the destructor also ends the lifetime of
`this`, and future attempts to use the field become an error when MSan's
use-after-dtor check is enabled.

The fix is to move the cleanup logic into a private helper function and
invoke the helper from both the destructor and the move assignment
operator.

Bug: 40222690
Change-Id: I1c31a1f2c342b0768ecc02f4f84c63b7e3a9160c
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/196554
Reviewed-by: Loko Kung <lokokung@google.com>
Commit-Queue: Daniel Cheng <dcheng@chromium.org>