Roll CTS: Rename *.worker.js to *.as_worker.js
This is a manual roll to a branch so that we can roll just the one
commit along with the necessary GN changes.

The next automated roll should switch back from this branch to the main
branch which already has this commit merged.

https://chromium.googlesource.com/external/github.com/gpuweb/cts/+log/fe3a24a38abc..b44989765d72

$ git log fe3a24a38..b44989765 --date=short --no-merges --format='%ad %ae %s'
2024-07-03 kainino Rename *.worker.js to *.as_worker.js (#3835)

Created with:
  roll-dep third_party/webgpu-cts

Bug: https://github.com/gpuweb/cts/issues/3834
Change-Id: Ie9cd4a6c2e1e3959967f2b9dbffc208354f92b6b
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/196874
Reviewed-by: Brian Sheedy <bsheedy@google.com>
Commit-Queue: Kai Ninomiya <kainino@chromium.org>
Reviewed-by: dan sinclair <dsinclair@chromium.org>