[hlsl] Support let pointer to var.
This Cl adds support to the DecomposeMemoryAccess transform to handle a
let which holds a var directly. The `var` is sunk down into any usages
of the `let` and the `let` removed. As similar thing happens to `let`
which directly holds an `access` to one of the `var`s. The `access` is
sunk down through the `let`.

Bug: 349867642
Change-Id: Ia6a090fc01e1b45510751f0d137f634bbd880f2d
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/196694
Reviewed-by: James Price <jrprice@google.com>
Commit-Queue: dan sinclair <dsinclair@chromium.org>