[hlsl] Add DecomposeMemoryAccess tests.
This Cl adds tests for the `uniform`, `workgroup`, `atomic` and store
requirements for the DecomposeMemoryAccess transform. Most of the new
tests are currently DISABLED.

Note, the result IR and HLSL is either a best guess or from the AST
transform. It's a guide as to what it will probably look like, but not
the final output.

Bug: 349867642
Change-Id: Ibacf359389d16d1c5790bbc565f5445a181aaef0
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/196955
Reviewed-by: James Price <jrprice@google.com>
Commit-Queue: dan sinclair <dsinclair@chromium.org>
Auto-Submit: dan sinclair <dsinclair@chromium.org>