[hlsl] Fixup loading of f16 vectors.
When loading an F16 vector in HLSL we have to make sure not to emit `>>`
which parses incorrect, instead we need to emit `> >`.

When templating the load, the `Load` instruction is always used, not the
`Load{n}` version for the size of the loaded vector.

Bug: 42251045
Change-Id: I3334ba26eb8a66b7e78c7db6fbccff21c091612e
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/196958
Reviewed-by: James Price <jrprice@google.com>
Commit-Queue: dan sinclair <dsinclair@chromium.org>