Add synchronous debug functionality for X11.
X11 errors can be very difficult to debug because of the asynchronous
nature of the protocol. Add a helper that makes all X11 calls
synchronous and triggers a breakpoint if an error is detect, such that
looking at the backtrace gives hints as to what went wrong.

Bug: None
Change-Id: I5e92878a4cd62a426a8faf81beecbd596c47fcad
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/189703
Reviewed-by: Stephen White <senorblanco@chromium.org>
Commit-Queue: Corentin Wallez <cwallez@chromium.org>