Move RefCountedWithExternalCount to dawn/common.
This is to make it possible to use in dawn/wire as well.

Also remove the convenience alias that implicitly use RefCounted as the
base class as dawn::wire won't do that (so only 2 out of 4 usages will
use RefCounted as the base class).

Bug: 344963953
Change-Id: I68c2bf8705452ae382f8f1349b47bf5bf75852b4
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/197194
Auto-Submit: Corentin Wallez <cwallez@chromium.org>
Commit-Queue: Loko Kung <lokokung@google.com>
Reviewed-by: Loko Kung <lokokung@google.com>
Reviewed-by: Austin Eng <enga@chromium.org>