Remove unused "ValidateVideoTextureCanBeShared"
ValidateVideoTextureCanBeShared() is unused in current dawn.
According to discussion in https://dawn-review.googlesource.com/c/dawn/+/174722:

"Import of multiplanar DXGI HANDLE also does not need a feature exposed because assuming the caller passes a valid HANDLE, then the caller needs to have successfully created that HANDLE. That means the caller needs to have successfully exported that HANDLE - which would not be possible if the device did not have sharedResourceTier2. In short, Dawn doesn't need the feature for import operations. We would need to expose such a feature if we allowed export of multiplanar share handles, but we don't have export today."

Bug: 351953347
Change-Id: Ie41be1e99587d298b4361a2d9d5bfab00c52ca07
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/197414
Commit-Queue: Shaobo Yan <shaobo.yan@intel.com>
Reviewed-by: Corentin Wallez <cwallez@chromium.org>