dawn::wire::server: Clear the object data when unregistering.
For most objects this does nothing as the ObjectData is plain old data.
However the data for Buffers contains Read/WriteHandles that otherwise
wouldn't be cleared until that slot is reused by a buffer.

Change the WireMemoryTransferServiceTests to explicitly destroy the
buffers to better control when the server read/write handles are
destroyed.

Bug: dawn:344963953
Change-Id: I39c21c4b9f74eb8fef0c2fa6d4d3de2458d1ccd7
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/197234
Reviewed-by: Loko Kung <lokokung@google.com>
Reviewed-by: Austin Eng <enga@chromium.org>
Commit-Queue: Corentin Wallez <cwallez@chromium.org>