generator_lib.py: Add support for importing templates in templates
Previously importing template with {% include %} or {% from %} would
cause issue with the integration with the build system because
generator_lib.py wouldn't know which additional template files to add as
dependencies to the generator task.

Change get_file_renders into get_outputs that return both a list of
FileRenders and a list of templates that will be imported. Use that new
list both in the dependency information of the build systems, and as an
allow list in the PreprocessingLoader (so we don't forget to add new
imported templates in the future).

Fixed: 352690884
Change-Id: I9d63a209d8c82405dec16c8f5baf8b1538cf2760
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/198254
Reviewed-by: Austin Eng <enga@chromium.org>
Commit-Queue: Corentin Wallez <cwallez@chromium.org>
Reviewed-by: Loko Kung <lokokung@google.com>