[YCbCr Samplers] Hold YCbCr desc in SamplerBase rather than SamplerVk
We don't currently take YCbCr sampler info into account when computing
the content hash of a Sampler or comparing Samplers for equality.

To prepare for fixing that bug, this CL changes the YCbCr desc to be
held as an ivar in SamplerBase rather than SamplerVk. This paves the
road for the next CL, which will add YCbCr info to the above
computations.

Change-Id: I7a666719851b691808bae9da0f5941321b7d871e
Bug: 41488897
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/196514
Commit-Queue: Colin Blundell <blundell@chromium.org>
Reviewed-by: Loko Kung <lokokung@google.com>