[Kotlin]: Add a KotlinRecord concept, factor struct conversion with it
In the Dawn code generators, both structures and methods are considered
"records" that have a bunch of named parameter/members with types. The
conversion of structures and method records from Kotlin to C++ are very
similar. To merge them both, a KotlinRecord concept is added that is a
C++ structure for a the JNI objects representing the values of the
Kotlin records.

A macro is added that creates the KotlinRecord structure for a record
(either a structure, or a method).

A macro is added that creates a conversion function from a KotlinRecord
to a structure.

These two macros are used to factor out all of the common logic out of
structures.cpp. The logic in methods.cpp will be factored out in a
follow-up CL.

A few new primitive types are added to dawn_kotlin.json.

Helper CallGetter functions are added in structures.cpp to call the
correct JNIEnv::Call*Method depending on what return type we expect.

A couple hacks for handling the jlong ANativeWindow are added. We should
remove them with a manually written method.

Bug: 352711433
Change-Id: Iaa89c9e7e85f724aa3fd98136acbad176f68e522
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/198236
Reviewed-by: Sonakshi Saxena <nexa@google.com>
Reviewed-by: Jim Blackler <jimblackler@google.com>
Commit-Queue: Corentin Wallez <cwallez@chromium.org>