[glsl] Add glsl validation
Run the glsl validator during the IR printer unittests. This requires
the code to be executable from an entry point (named `main`). The type
of the entry point must be provided to the `Generate` method (defaults
to `compute` to match the empty entry point transform).

Bug: 42251044
Change-Id: I303fc0fe170cde398870479bfeda594776468574
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/200456
Reviewed-by: James Price <jrprice@google.com>
Commit-Queue: dan sinclair <dsinclair@chromium.org>