[ir] Refactor type validation
Move the check for reference types to a common `CheckType()`, so that
we can easily add other types to check for in the future. Also add a
call to this function for block parameters, which are the other place
that a type can show up.

Some error messages become slightly less detailed, but not in a
significant way.

Change-Id: I2a7dca3e866abe02b4b86a5b0b9c464ee72fdfa2
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/199855
Reviewed-by: Ryan Harrison <rharrison@chromium.org>
Reviewed-by: dan sinclair <dsinclair@chromium.org>
Commit-Queue: James Price <jrprice@google.com>