[tint] Fix undef arg after ArrayLengthFromUniform
When a function calls `arrayLength` with a user function parameter, it
always expects to receive the length via another parameter. If the
originating variable at the callsite is not in the bindpoint map, we
do not retrieve the length from a UBO, which led to `undef` being
passed as the length argument.

In this scenario, the caller should just call the original
`arrayLength` builtin to get the length.

Fixed: 356441926
Change-Id: I458dfd848c8b4633976ab080e6fd940f1e7be992
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/200920
Auto-Submit: James Price <jrprice@google.com>
Commit-Queue: James Price <jrprice@google.com>
Reviewed-by: dan sinclair <dsinclair@chromium.org>
Commit-Queue: dan sinclair <dsinclair@chromium.org>