[ir] Validate nested pointers
Reject nested pointers and references by default. Fix a small bug in
the type traversal code that was preventing it from looking at pointer
store types.

The MSL backend does use pointers inside structures, so add a
capability for that and enable it in transform that need to consume
modules that do this.

Change-Id: I09a3ba0c46a18e2477d79710155075eef1188b93
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/199856
Commit-Queue: James Price <jrprice@google.com>
Reviewed-by: Ryan Harrison <rharrison@chromium.org>
Reviewed-by: dan sinclair <dsinclair@chromium.org>