Partial MSAA expand-resolve
This adds a 'dawn_partial_load_resolve_texture' feature extending
'dawn_load_resolve_texture' to support partial MSAA expand-resolve.
'RenderPassDescriptorExpandResolveRect' can be chained into
'RenderPassDescriptor' to indicate that only the region of rect needs to
be expanded and resolved, so that we don't need to do that always
against the whole MSAA texture.

Bug: chromium:344814092
Change-Id: I33fd40e4ea7bae8e134eef6a2b9a7f2b13cf949d
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/199314
Reviewed-by: Quyen Le <lehoangquyen@chromium.org>
Commit-Queue: Jie A Chen <jie.a.chen@intel.com>
Reviewed-by: Corentin Wallez <cwallez@chromium.org>