Add a sorted `UsagesSorted`.
This CL adds a `UsagesSorted` call to the ir::Value which returns a
sorted list of usages.

The usages are sorted by `<instruction id, operand index>`. In the case
of a `nullptr` instruction the null usages come last and are then sorted
by operand index.

Bug: 354711610
Change-Id: I6f94c3d6a8d8d5ffd6fdde8fa6b6ffe02a6d2a08
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/201015
Reviewed-by: Antonio Maiorano <amaiorano@google.com>
Commit-Queue: dan sinclair <dsinclair@chromium.org>