[tint] Implement subgroupShuffle
Implement subgroupShuffle builtin for WGSL, SPIR-V IR backend (not AST),
MSL AST and IR backends and HLSL AST and IR backends. Subgroup builtins
are not yet implemented for the GLSL backend.

A polyfill is necessary in the SPIR-V IR backend to handle the case
where the lane ID is a signed integer.

Bug: 354738715
Change-Id: I3c29c1862d29600457fdb790b954fe0854008b52
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/201956
Reviewed-by: dan sinclair <dsinclair@chromium.org>
Commit-Queue: Natalie Chouinard <chouinard@google.com>