[tint][spirv] Allow signed subgroupBroadcast ID
Allow the id argument for the subgroupBroadcast built-in function to be
i32 or u32. This requires compile-time conversion to u32 when the id is
const i32. There is no change to the requirement that this parameter is
const yet.

This fix has only been applied to the SPIR-V IR backend.

Resolves failing CTS test:
webgpu:shader,validation,expression,call,builtin,subgroupBroadcast:id_type:type="i32"

Bug: 354224075
Change-Id: I2b3b83f45db1b5c048d8fc2d8200f687ffa9da60
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/202074
Reviewed-by: dan sinclair <dsinclair@google.com>
Reviewed-by: James Price <jrprice@google.com>
Commit-Queue: Natalie Chouinard <chouinard@google.com>